requirements.txt
mcp
openstackclient
